+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.5.0 (unreleased)
+## 0.5.1 (unreleased)
+
+IMPROVEMENTS:
+
+  * provider/aws: `aws_s3_bucket` exports `hosted_zone_id` and `region` [GH-1865]
+  * provider/aws: `aws_route53_record` exports `fqdn` [GH-1847]
+
+BUG FIXES:
+
+  * provider/aws: Mark `aws_security_group` description as `ForceNew` [GH-1871]
+
+## 0.5.0 (May 7, 2015)
 
 FEATURES:
 
@@ -10,6 +21,13 @@ FEATURES:
      `TF_VAR_name` and this will be checked last for a value.
   * **New remote state backend: `s3`**: You can now store remote state in
      an S3 bucket. [GH-1723]
+  * **Automatic AWS retries**: This release includes a lot of improvement
+     around automatic retries of transient errors in AWS. The number of
+     retry attempts is also configurable.
+  * **Templates**: A new `template_file` resource allows long strings needing
+     variable interpolation to be moved into files. [GH-1778]
+  * **Provision with WinRM**: Provisioners can now run remote commands on
+     Windows hosts. [GH-1483]
 
 IMPROVEMENTS:
 
@@ -18,9 +36,24 @@ IMPROVEMENTS:
   * **New resource: `aws_app_cookie_stickiness_policy`**
   * **New resource: `aws_customer_gateway`**
   * **New resource: `aws_ebs_volume`**
+  * **New resource: `aws_elasticache_cluster`**
+  * **New resource: `aws_elasticache_security_group`**
+  * **New resource: `aws_elasticache_subnet_group`**
+  * **New resource: `aws_iam_access_key`**
+  * **New resource: `aws_iam_group_policy`**
+  * **New resource: `aws_iam_group`**
+  * **New resource: `aws_iam_instance_profile`**
+  * **New resource: `aws_iam_policy`**
+  * **New resource: `aws_iam_role_policy`**
+  * **New resource: `aws_iam_role`**
+  * **New resource: `aws_iam_user_policy`**
+  * **New resource: `aws_iam_user`**
   * **New resource: `aws_lb_cookie_stickiness_policy`**
-  * **New resource: `aws_vpc_dhcp_options`**
+  * **New resource: `aws_proxy_protocol_policy`**
+  * **New resource: `aws_security_group_rule`**
   * **New resource: `aws_vpc_dhcp_options_association`**
+  * **New resource: `aws_vpc_dhcp_options`**
+  * **New resource: `aws_vpn_connection_route`**
   * **New resource: `google_dns_managed_zone`**
   * **New resource: `google_dns_record_set`**
   * **Migrate to upstream AWS SDK:** Migrate the AWS provider to
@@ -30,6 +63,8 @@ IMPROVEMENTS:
       migrated back to the upstream version.
   * core: Improve error message on diff mismatch [GH-1501]
   * provisioner/file: expand `~` in source path [GH-1569]
+  * provider/aws: Better retry logic, now retries up to 11 times by default
+      with exponentional backoff. This number is configurable. [GH-1787]
   * provider/aws: Improved credential detection [GH-1470]
   * provider/aws: Can specify a `token` via the config file [GH-1601]
   * provider/aws: Added new `vpc_security_group_ids` attribute for AWS
@@ -37,13 +72,24 @@ IMPROVEMENTS:
       Instance without destroying it [GH-1539]
   * provider/aws: White or blacklist account IDs that can be used to
       protect against accidents. [GH-1595]
+  * provider/aws: Add a subset of IAM resources [GH-939]
+  * provider/aws: `aws_autoscaling_group` retries deletes through "in progress"
+      errors [GH-1840]
+  * provider/aws: `aws_autoscaling_group` waits for healthy capacity during
+      ASG creation [GH-1839]
   * provider/aws: `aws_instance` supports placement groups [GH-1358]
+  * provider/aws: `aws_eip` supports network interface attachment [GH-1681]
   * provider/aws: `aws_elb` supports in-place changing of listeners [GH-1619]
   * provider/aws: `aws_elb` supports connection draining settings [GH-1502]
   * provider/aws: `aws_elb` increase default idle timeout to 60s [GH-1646]
   * provider/aws: `aws_key_pair` name can be omitted and generated [GH-1751]
+  * provider/aws: `aws_network_acl` improved validation for network ACL ports
+      and protocols [GH-1798] [GH-1808]
+  * provider/aws: `aws_route_table` can target network interfaces [GH-968]
+  * provider/aws: `aws_route_table` can specify propogating VGWs [GH-1516]
   * provider/aws: `aws_route53_record` supports weighted sets [GH-1578]
   * provider/aws: `aws_route53_zone` exports nameservers [GH-1525]
+  * provider/aws: `aws_s3_bucket` website support [GH-1738]
   * provider/aws: `aws_security_group` name becomes optional and can be
       automatically set to a unique identifier; this helps with
       `create_before_destroy` scenarios [GH-1632]
@@ -52,6 +98,9 @@ IMPROVEMENTS:
   * provider/aws: automatically set the private IP as the SSH address
       if not specified and no public IP is available [GH-1623]
   * provider/aws: `aws_elb` exports `source_security_group` field [GH-1708]
+  * provider/aws: `aws_route53_record` supports alias targeting [GH-1775]
+  * provider/aws: Remove default AWS egress rule for newly created Security Groups [GH-1765]
+  * provider/consul: add `scheme` configuration argument [GH-1838]
   * provider/docker: `docker_container` can specify links [GH-1564]
   * provider/google: `resource_compute_disk` supports snapshots [GH-1426]
   * provider/google: `resource_compute_instance` supports specifying the
@@ -61,6 +110,7 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+  * core: Fix graph cycle issues surrounding modules [GH-1582] [GH-1637]
   * core: math on arbitrary variables works if first operand isn't a
       numeric primitive. [GH-1381]
   * core: avoid unnecessary cycles by pruning tainted destroys from
@@ -83,6 +133,8 @@ BUG FIXES:
       that would previously only show up during apply [GH-1655]
   * core: Referencing invalid module output in module validates [GH-1448]
   * command: remote states with uppercase types work [GH-1356]
+  * provider/aws: Support `AWS_SECURITY_TOKEN` env var again [GH-1785]
+  * provider/aws: Don't save "instance" for EIP if association fails [GH-1776]
   * provider/aws: launch configuration ID set after create success [GH-1518]
   * provider/aws: Fixed an issue with creating ELBs without any tags [GH-1580]
   * provider/aws: Fix issue in Security Groups with empty IPRanges [GH-1612]
@@ -92,10 +144,15 @@ BUG FIXES:
   * provider/aws: Block devices can be encrypted [GH-1718]
   * provider/aws: ASG health check grace period can be updated in-place [GH-1682]
   * provider/aws: ELB security groups can be updated in-place [GH-1662]
+  * provider/aws: `aws_main_route_table_association` can be deleted
+      manually [GH-1806]
+  * provider/docker: image can reference more complex image addresses,
+      such as with private repos with ports [GH-1818]
   * provider/openstack: region config is not required [GH-1441]
   * provider/openstack: `enable_dhcp` for networking subnet should be bool [GH-1741]
   * provisioner/remote-exec: add random number to uploaded script path so
       that parallel provisions work [GH-1588]
+  * provisioner/remote-exec: chmod the script to 0755 properly [GH-1796]
 
 ## 0.4.2 (April 10, 2015)